refactor: move files to internal dir
