+++ src/internal/token.rs
@@ -1,4 +1,4 @@
-use crate::token_type::TokenType;
+use crate::internal::token_type::TokenType;
 
 #[derive(Debug, Clone)]
 pub enum Literal {
